ClangToModel: accept plain C builtin types
`edit-c-type` and `edit-c-body` rejected every builtin spelled as a
plain C type (`int`, `void *`, ...).

This commit adds support for them by converting them upon import using
information about the target ABI.

+++ share/revng/test/tests/analysis/EditCTypeAnalysis/struct-with-an-unsupported-field/edit.c
@@ -4,5 +4,5 @@
 
 struct _PACKED _SIZE(24) my_struct {
   _STARTS_AT(16)
-  int unsupported_field;
+  _Float16 unsupported_field;
 };
